qcacld-3.0: Set link status to 'discovering' based on TDLS support
When TDLS peer capabilities are unknown, the TDLS peer state does not
change to discovering. So even when we receive a discovery response
frame, the setup is not initiated.

Instead of checking if the peer TDLS capability is unknown,
check if the peer doesn't support TDLS. If the peer doesn't support
TDLS, don't change the TDLS state to discovering.

CRs-Fixed: 4052954
Change-Id: Icf673edf5ebd58678b080d57b44b4d4870c6d649