disp: msm: clear DMS flag when active changed is set
HAL compositor can select any of the possible crtcs
for queuing first commit on bootup. This is causing the
failures in DSI with DMS flag populated and the pm runtime
votes are not being removed.

Add changes to remove the DMS flag on active changed
to a new placeholder in the _dsi_bridge_mode_validate_and_fixup
api.

Change-Id: Ic0ac740c39c2ffb8868fd342dfb062ef9f0a62ae
Signed-off-by: Rehan Kumar <quic_rehakuma@quicinc.com>